added:adding and showing todo list,status handler, refactor:handler code

+++ public/js/todoBox.js
@@ -29,21 +29,31 @@ const getDateString = function(dateOb) {
   return `${month} ${date} ${day}`;
 };
 
-const getEachItemHtml = function(item) {
-  return `<div class="eachItem" id="${item.id}">
-  <div class="cardUntick"></div>
-  <div class="itemName">${item.name}</div>
-  <div class="deleteItem">X</div>
-</div>`;
+const getEachItemHtml = function(ext, deleteEnable, item) {
+  const html = `<div class="${ext}eachItem" id="${item.id}">
+  <div class="${ext}cardUntick">
+  <div class="${ext}cardTick"></div>
+  </div>
+  <div class="${ext}itemName">${item.name}</div>`;
+  if (deleteEnable) {
+    return `${html}
+    <div class="deleteItem">X</div>
+    </div>`;
+  }
+  return `${html}</div>`;
 };
 
-const getListHtml = function(items) {
-  return items.map(getEachItemHtml).join('');
+const getListHtml = function(list, mapper) {
+  return list.map(mapper).join('');
 };
 
 const todoBox = function(res) {
   const date = getDateString(new Date(res.date));
-  const listHtml = getListHtml(res.items);
+  const listHtml = getListHtml(
+    res.items,
+    getEachItemHtml.bind(null, '', true)
+  );
+
   return `<div class="heading">${res.title}</div>
       <div class="date">${date}</div>
       <div class="smallHorizontalLine"></div>
@@ -58,3 +68,25 @@ const todoBox = function(res) {
 const titleHtml = `<div class="writeTitle">write Title</div>
 <input type="text" id="titleInput" />
 <button id="done" onclick="takeTitle()">Done</button>`;
+
+const getEachTodoHtml = function(todo) {
+  const items = getListHtml(
+    todo.items,
+    getEachItemHtml.bind(null, 'todo-', false)
+  );
+  return `<div class="cardBox" id="${todo.id}">
+   <div class="todoCard">
+     <div class="todoCardTitle">${todo.title}</div>
+     ${items}
+   </div>
+   <div class="topTick"></div>
+ </div>`;
+};
+
+const todoCards = function(res) {
+  let div = '';
+  for (const todo in res) {
+    div += getEachTodoHtml(res[todo]);
+  }
+  return div;
+};
